Implement create recipe

+++ RecipeApp.Domain/Entities/Models/Instruction.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public int RecipeId { get; set; }
-        public int InstructionNumber { get; set; }
+        public byte InstructionNumber { get; set; }
         public string Description { get; set; } = null!;
         public virtual Recipe Recipe { get; set; }
     }
